airflow_ml_dags/requirements.txt
apache-airflow
scikit-learn
numpy
pytest
dag_constants
